autogen/requirements.txt
pyautogen[teachable]
pymemgpt